Create saiyan boss logic (#15)
* switched mobs to use abilities like player

-- work in progress, needs to check KI

* Interrupt might be working but needs tweaking

Task can be interrupted but have a bug with the default animation and the mob dying

* added cancel to kick

* energy barrage can now be canceled

* all beams should now be cancellable

- added KI bar to mobs so we can visualise the regen

* added first draft of saiyan boss

* Boss now works with multiple abilities

* added Cui as a playable character

* updated path to spriteframes for Cui

* starting new spawn rules

* Implementadas as Spawn Rules de 0 a 180 s (até 3min)

* fix: ability manager error fix, and kamehameha cancelation of sound.

* minor tweaks, galick gun boss as a major issue that breaks the game, when in phase 2, cant figure it out.

* small tweak on the process of the mob to make the knockback work.

* kick working with normal mobs again and other small tweaks

---------

+++ Scenes/Abilities/Ability.cs
@@ -43,9 +43,9 @@ public partial class Ability : Node2D
 
   public virtual void Cancel() { GD.Print("Canceling ability: " + Name); }
 
-  public virtual void SpendKi(int value)
+  public virtual void SpendKi(float value)
   {
     EventRegistry.GetEventPublisher("IncreaseStatsFromDictionary")
-      .RaiseEvent(new object[] { new Dictionary<string, int> { { "ki", -abilityResource.kiRequired } }, this });
+      .RaiseEvent(new object[] { new Dictionary<string, float> { { "ki", -abilityResource.kiRequired } }, this });
   }
 }
